EncryptoVision/requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
Bottleneck @ file:///D:/bld/bottleneck_1602497448200/work
cachetools==5.3.1
certifi==2022.12.7
cffi @ file:///D:/bld/cffi_1671179514672/work
charset-normalizer==2.1.1
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
contourpy==1.2.1
cryptography @ file:///D:/bld/cryptography-split_1685659723888/work
cycler==0.12.1
datasets==2.18.0
decorator==4.4.2
dgl==1.1.1+cu118
dill==0.3.8
dnspython @ file:///home/conda/feedstock_root/build_artifacts/dnspython_1673812078720/work
docker-pycreds==0.4.0
et-xmlfile==1.1.0
filelock==3.9.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
future @ file:///home/conda/feedstock_root/build_artifacts/future_1673596611778/work
fvcore==0.1.5.post20221221
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.21.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.11.0
huggingface-hub==0.22.2
hyperopt==0.1.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
importlib-metadata==6.6.0
importlib_resources==6.4.0
iopath==0.1.10
ipaddress==1.0.23
Jinja2==3.1.2
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1663332044897/work
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.2.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.5.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.2.1
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.7
networkx==2.1
numexpr @ file:///D:/bld/numexpr_1602669914655/work
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas @ file:///C:/b/abs_cdcgk91igc/croots/recipe/pandas_1663772960432/work
Pillow==9.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1687705014305/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1679580333621/work
portalocker==2.8.2
protobuf==3.20.3
psutil @ file:///D:/bld/psutil_1681775298734/work
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydeck==0.8.1b0
Pygments==2.17.2
pymongo @ file:///D:/bld/pymongo_1687693431930/work
pyparsing==3.1.2
PySocks @ file:///D:/bld/pysocks_1661604991356/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1680088766131/work
pywin32==306
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scapy==2.5.0
scikit-learn @ file:///D:/bld/scikit-learn_1604232709065/work
scikit-plot==0.3.7
scipy @ file:///C:/bld/scipy_1604305141961/work
sentencepiece==0.2.0
sentry-sdk==1.45.0
seqeval==1.2.2
setproctitle==1.3.3
setuptools-scm==8.1.0
six==1.15.0
smmap==5.0.1
some-package==0.1
streamlit==1.33.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==1.1.0
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1643647933166/work
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.0.1+cu118
torch_geometric==2.5.2
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.4
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1677948868469/work
transformers==4.39.3
typeguard==2.13.3
typing_extensions==4.10.0
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1680081134351/work
urllib3==1.26.13
wandb==0.16.6
watchdog==4.0.0
Werkzeug==2.3.6
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
wrapt==1.12.1
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
zipp==3.15.0
